+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.0.59
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.32.4
+
 ## 0.0.58
 
 ### Patch Changes